[flutter_web] move CellView to lib/views
